Add a workstation::r10k class
This is an odd duck.  Need it to install r10k into our install puppet-agent
because we have no system ruby yet.  This class will be applied by the
bootstrap script so that we can then r10k, so that we then have the modules
we need to actually puppet apply workstation state...