(maint) Add a type for a repository
and a single class for managing repositories with the repos/repo defined
types. Currently this works by consuming a large Array of hashes data
structure which maps down to workstation::repos parameters. This could
be changed at some point to be a custom manifest if I introduce a way to
add custom manifests instead of maintaining the definition in hiera.

Also adds a type for defining remotes to be added to cloned
repositories.

And specs for the repo types/classes.