(maint) Clone dotfiles to a bare git repository
and checkout into $HOME. Replaces the execution of a custom sync script
from dotfiles to instead just checkout with aliased git and work dirs
so as to avoid other repos in $HOME being nested within the base
dotfiles checkout.

This matches up with changes to the dotfiles structure to expect being
checked out into $HOME instead of symlinked.

Pass a flag through to vcsrepo to ensure a bare git clone if specified.