(maint) Optionally check out a branch when setting up a repo
If a branch is checked out, it will be set to either the upstream or
origin branch. In the case of a branch, such as master, that is likely
to already exist in origin but be out of date of upstream, it will sync
to upstream and pull it up to date.