(maint) Fix repo fetch/pull for upstream/origin remotes
When the bin/install.sh script is executed, we're shelling into the host
to be managed as the default container user ('ubuntu'), but we are
preparing repositories as the dev user ('jpartlow' in my case). Since
this involves interaction with github, ssh keys are involved, and we're
relying on the ssh-agent being forwarded.

This works fine when ubuntu sudo's as root so long as the environment is
passed along, but when root attempts to shift down to jpartlow to
perform the pull/fetch operations, it can no longer reach the ssh-agent
socket which is owned by ubuntu.

To get around this, we're skipping 'user => jpartlow' in the Execs for
these actions, which means root is making the requests and we have
access to the SSH_AUTH_SOCK.

But the chown -R jpartlow:jpartlow repo cleanup is required in these
cases. (This is what the vcsrepo module ends up doing as well in its
provider.)

There are some other fixes in this patch to ensure that execs remain
unique and that ordering is correct.