(maint) Use Bolt
Reworks the module to use puppet-bolt, which let's me ditch the hacky
bin/install.sh which was providing the remote execution bootstrap, and
also the additional step of checking out the module on the remote host.

Now that we have Bolt it provides a much cleaner process for all of
this, the new workstation::manage plan (in plans/manage.pp) uses
apply_prep to get puppet-agent installed on the node, and then can apply
'include workstation' building the catalog locally (after a `bolt
puppetfile install has installed dependencies`), and then shipping it to
remote workstation and applying it there.

It still depends on Hiera data during the catalog
compilation to provide the correct configuration for the module, and
requries that the vm were managing can be reached via ssh with agent
forwarding enabled.