(maint) Manage bolt via the puppet-tools-release repository
Recently bolt was no longer shipped in the puppet-release repository
that it added automatically to the node by apply_prep (when it installs
puppet). Now bolt is in the puppet-tools-release repository that needs
to be installed separately.

Made use of the previous commits changes to repository_packages to get
update both plans to manage bolt correctly.