(maint) Display errors as well as Puppet logs
When an apply fails, if there was a catalog compilation failure it was
being missed, because the function would instead raise an error on
$report['logs'] because $report would be undef.

Ensures, instead, that we don't fail for that case and do raise the
actual apply compilation failure.

If you were managing multiple nodes, this would interrupt running on
all, which is possibly an issue.