(maint) Simplify setup_test_tools plan by using dev_account_base
...profile which manages a set of the classes that this plan was
including.

Further work needs to be done to make classes for the nfs and any other
bolt tool workstation specific management. This plan should be dropped,
and it should be possible just to run the manage plan with a
workstation::profiles list. The use of upload_file() for the the
vmfloaty and control repo key stand in the way of that, but I don't
think's necessary to use upload_file there.