(maint) Manage subdirs of repository paths automatically
The manifests generate git repos under working paths defined in the
repository data hash. Ensuring that those subdirectories existed was an
external exercise, and there were explicit file references in the
workstation class, for example ensuring my defaults.

For the more general case of using the 'manage' plan with some set of
profile classes like dev_account_base, I don't know ahead of time what
the repo paths will be. The defined type added here ensures the path
subdirectories, relative to a root prefix are created. Since the
repository data structure may contain multiple entries with the same
paths, it only manages the directories if a file resource does not
already exist.

This doesn't handle the case of creating a file resource for one of the
dirs in a later class and hitting a duplicate resource.